# project/web-go/docs/swagger.yaml
basePath: /api/v1
definitions:
  sudooom_im_web_internal_service.FriendRequestRequest:
    properties:
      friend_id:
        type: integer
      message:
        type: string
    required:
    - friend_id
    type: object
  sudooom_im_web_internal_service.LoginRequest:
    properties:
      device_id:
        type: string
      password:
        type: string
      platform:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  sudooom_im_web_internal_service.LoginResponse:
    properties:
      access_token:
        type: string
      expires_at:
        type: integer
      object_code:
        type: string
      refresh_token:
        type: string
      user_id:
        type: integer
    type: object
  sudooom_im_web_internal_service.RegisterRequest:
    properties:
      nickname:
        maxLength: 50
        minLength: 1
        type: string
      password:
        maxLength: 50
        minLength: 6
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - nickname
    - password
    - username
    type: object
  sudooom_im_web_internal_service.UpdateProfileRequest:
    properties:
      avatar:
        type: string
      nickname:
        type: string
    type: object
  sudooom_im_web_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: IM 即时通讯 Web 服务 API
  title: IM Web API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: 用户登录获取 Token
      parameters:
      - description: 登录信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sudooom_im_web_internal_service.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sudooom_im_web_pkg_response.Response'
      summary: 用户登录
      tags:
      - 认证
  /auth/logout:
    post:
      consumes:
      - application/json
      description: 用户登出，Token 失效
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sudooom_im_web_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 用户登出
      tags:
      - 认证
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: 使用 refresh_token 获取新的 access_token
      parameters:
      - description: 刷新 Token
        in: body
        name: request
        required: true
        schema:
          properties:
            refresh_token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sudooom_im_web_pkg_response.Response'
      summary: 刷新 Token
      tags:
      - 认证
  /auth/register:
    post:
      consumes:
      - application/json
      description: 创建新用户账号
      parameters:
      - description: 注册信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sudooom_im_web_internal_service.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sudooom_im_web_pkg_response.Response'
      summary: 用户注册
      tags:
      - 认证
  /friends:
    get:
      description: 获取当前用户的好友列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sudooom_im_web_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取好友列表
      tags:
      - 好友
  /friends/{id}:
    delete:
      description: 删除指定的好友关系
      parameters:
      - description: 好友用户 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sudooom_im_web_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 删除好友
      tags:
      - 好友
  /friends/accept/{id}:
    post:
      description: 接受指定的好友请求
      parameters:
      - description: 好友请求 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sudooom_im_web_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 接受好友请求
      tags:
      - 好友
  /friends/reject/{id}:
    post:
      description: 拒绝指定的好友请求
      parameters:
      - description: 好友请求 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sudooom_im_web_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 拒绝好友请求
      tags:
      - 好友
  /friends/request:
    post:
      consumes:
      - application/json
      description: 向指定用户发送好友请求
      parameters:
      - description: 好友请求信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sudooom_im_web_internal_service.FriendRequestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sudooom_im_web_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 发送好友请求
      tags:
      - 好友
  /friends/requests:
    get:
      description: 获取发送给当前用户的待处理好友请求
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sudooom_im_web_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取待处理的好友请求
      tags:
      - 好友
  /user/{id}:
    get:
      description: 通过用户 ID 获取用户信息
      parameters:
      - description: 用户 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sudooom_im_web_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取指定用户信息
      tags:
      - 用户
  /user/profile:
    get:
      description: 获取当前登录用户的详细信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sudooom_im_web_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取当前用户信息
      tags:
      - 用户
    put:
      consumes:
      - application/json
      description: 更新当前登录用户的信息
      parameters:
      - description: 更新信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sudooom_im_web_internal_service.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sudooom_im_web_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 更新用户信息
      tags:
      - 用户
  /user/search:
    get:
      description: 根据关键词搜索用户
      parameters:
      - description: 搜索关键词
        in: query
        name: keyword
        required: true
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sudooom_im_web_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 搜索用户
      tags:
      - 用户
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"